#369 Move as many functions as possible into source files for the lib
Fix #369
