+++ cmake/build_types/library.cmake
@@ -2,16 +2,23 @@
 ### Distributed under the Boost Software License, Version 1.0.
 ### (See accompanying file LICENSE or copy at https://www.boost.org/LICENSE_1_0.txt)
 
-add_library(arg_router INTERFACE ${HEADERS})
-target_link_libraries(arg_router INTERFACE
+set(LIBRARY_TYPE STATIC)
+if (NOT BUILD_STATIC)
+    set(LIBRARY_TYPE SHARED)
+endif()
+
+add_library(arg_router ${LIBRARY_TYPE} ${HEADERS} ${SOURCES})
+
+set_target_properties(arg_router PROPERTIES CXX_EXTENSIONS OFF)
+target_compile_features(arg_router PUBLIC cxx_std_20)
+
+target_link_libraries(arg_router PUBLIC
     Boost::boost
 )
 
-target_include_directories(arg_router AFTER INTERFACE
+target_include_directories(arg_router AFTER PUBLIC
     $<BUILD_INTERFACE:${CMAKE_SOURCE_DIR}/include>
     $<INSTALL_INTERFACE:include>  # <prefix>/include
 )
 
-if(NOT INSTALLATION_ONLY)
-    add_dependencies(arg_router clangformat)
-endif()
+add_dependencies(arg_router clangformat)
